+++ src/Credentials/ServiceAccountCredentials.php
@@ -22,6 +22,7 @@ use Google\Auth\GetQuotaProjectInterface;
 use Google\Auth\OAuth2;
 use Google\Auth\ServiceAccountSignerTrait;
 use Google\Auth\SignBlobInterface;
+use InvalidArgumentException;
 
 /**
  * ServiceAccountCredentials supports authorization using a Google service
@@ -81,11 +82,13 @@ class ServiceAccountCredentials extends CredentialsLoader implements SignBlobInt
      *   as an associative array
      * @param string $sub an email address account to impersonate, in situations when
      *   the service account has been delegated domain wide access.
+     * @param string $targetAudience The audience for the ID token.
      */
     public function __construct(
         $scope,
         $jsonKey,
-        $sub = null
+        $sub = null,
+        $targetAudience = null
     ) {
         if (is_string($jsonKey)) {
             if (!file_exists($jsonKey)) {
@@ -107,6 +110,14 @@ class ServiceAccountCredentials extends CredentialsLoader implements SignBlobInt
         if (array_key_exists('quota_project', $jsonKey)) {
             $this->quotaProject = (string) $jsonKey['quota_project'];
         }
+        if ($scope && $targetAudience) {
+            throw new InvalidArgumentException(
+                'Scope and targetAudience cannot both be supplied');
+        }
+        $additionalClaims = [];
+        if ($targetAudience) {
+            $additionalClaims = ['target_audience' => $targetAudience];
+        }
         $this->auth = new OAuth2([
             'audience' => self::TOKEN_CREDENTIAL_URI,
             'issuer' => $jsonKey['client_email'],
@@ -115,6 +126,7 @@ class ServiceAccountCredentials extends CredentialsLoader implements SignBlobInt
             'signingKey' => $jsonKey['private_key'],
             'sub' => $sub,
             'tokenCredentialUri' => self::TOKEN_CREDENTIAL_URI,
+            'additionalClaims' => $additionalClaims,
         ]);
     }